% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plus.R
\name{plus}
\alias{plus}
\title{Sum of two numbers}
\usage{
plus(x, y)
}
\arguments{
\item{x}{first number}

\item{y}{second number}
}
\value{
sum of x and y
}
\description{
Sum of two numbers
}
\examples{
plus(4, 6)
}
